Change SQL connection information to be provided at runtime
The SQL root password is no longer contained within the jar or source
code.

Instead, spring is now set-up so that the URL, username, and password
for the MySQL credentials database are provided as system properties,
which are set via a jetty WebAppContext xml file, which are provided by
docker/k8s at container creation time.

This is a big hack and I couldn't find any simpler way to do this.